fix: update wrong dependencies and split HttpAdapter to a single file
