+++ connectors/source-http/pom.xml
@@ -16,8 +16,8 @@
 
         <dependency>
             <groupId>com.linkall</groupId>
-            <artifactId>vance-cdk</artifactId>
-            <version>1.0.0</version>
+            <artifactId>cdk-java</artifactId>
+            <version>0.2.0</version>
         </dependency>
 
     </dependencies>
